Rework the way that DBus listens for incoming messages
This patch also introduces an event loop for Skype. This is necessary
for DBus. You have two methods of running the loop: #run which doesn't
return naturally, and #tick which does a single run of the loop for
integration with other event loops such as a GTK loop or similar.